+++ ChangeLog.md
@@ -9,6 +9,7 @@ Blob:
 - Fixed issue of setting blob tag should not update Blob Etag and LastModified. (issue #2327)
 - Fix HTTP header parsing of `SubmitBatch()`. If a HTTP header has HTTP header delimiter (`:`) in its value, `SubmitBatch()` returns "400 One of the request inputs is not valid". For example, if `user-agent` header is `azsdk-cpp-storage-blobs/12.10.0-beta.1 (Darwin 23.1.0 arm64 Darwin Kernel Version 23.1.0: Mon Oct  9 21:28:12 PDT 2023; root:xnu-10002.41.9~6/RELEASE_ARM64_T8103)`, all `SubmitBatch()` requests are failed.
 - Fixed issue of blob copying succeed without 'r' permission in source blob's SAS token credential.
+- Fixed issue of list container contains metadata even request doesn't have include=metadata (issue #2382)
 
 ## 2023.12 Version 3.29.0